fix(app): isolate AppSettings tests from the production Keychain account
`testOpenAIAPIKeyViaKeychainHelper` drove the real Keychain slot that
`AppSettings.openAIAPIKey` reads and writes — service
`com.meetingtranscriber`, account `openAIAPIKey`. On any machine that
also runs the app, that has two consequences. The item is owned by the
app's code signature, so the test binary touching it raises an
authorization prompt and `swift test` blocks indefinitely, which is the
reported symptom. And if that prompt is granted, the unconditional
`delete` at the top of the test destroys the API key the user configured
in Settings.

Give `AppSettings` an injectable Keychain account defaulting to the
production one, mirroring the `defaults:` injection already there, and
have the fixture mint a unique account per test. Asserting through the
injected account keeps the test honest rather than merely quiet: if
`AppSettings` ignored the parameter, every read would come back nil and
the test fails. Verified by mutating the initializer to use a different
account — the test goes red on exactly that assertion.

`AppSettings.swift` sat at 599 of its 600-line `file_length` budget, so
the two added lines tripped `swiftlint --strict`. The existing `Computed`
section moves verbatim into `AppSettings+Computed.swift` to make room,
following the `AppSettings+RPC.swift` precedent. No behaviour change.

Closes #584
